Add pseudofree energy support to pfn opt

+++ src/api/bridge/rnastructure.cpp
@@ -69,7 +69,7 @@ struct PfnState {
 
   PfnState(int N, datatable* data)
       : w(N), v(N), wmb(N), wl(N), wlc(N), wmbl(N), wcoax(N), w5(new PFPRECISION[N + 1]),
-        w3(new PFPRECISION[N + 2]), pfdata(new pfdatatable(data, scaling, T)),
+        w3(new PFPRECISION[N + 2]), pfdata(new pfdatatable(data, scaling, double(T))),
         fce(new forceclass(N)), lfce(new bool[2 * N + 1]()), mod(new bool[2 * N + 1]()) {}
 };
